backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.31.0
greenlet==3.2.4
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.7
email-validator==2.1.0
bcrypt==4.0.1